in NarInfo parsing make values in References Node optional
* Sometimes I find narinfo files have an empty "References:" node (in
terms of YAML, though it doesn't have to be YAML). The parsing code from
`nix/src/libstore/nar-info.cc` does not treat empty references as an
error. So I do and I don't see problems with it.
* There is confusion with another situation (I find more often) when
narinfo references itself. Why is that? Or why is it not consistent
across all narinfos?